reorganised logic in addView - returned stuff from ViewModel
